+++ core/src/androidTest/java/de/raphaelebner/roomdatabasebackup/core/ExampleInstrumentedTest.kt
@@ -1,0 +1,22 @@
+package de.raphaelebner.roomdatabasebackup.core
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("com.ebner.roomdatabasebackup.core.test", appContext.packageName)
+    }
+}
